+++ api.h
@@ -9,6 +9,7 @@
 #include <utility>
 #include <map>
 #include <set>
+#include <cstdint> //has type 'uint64_t'
 
 using std::pair;
 using std::map;
@@ -17,14 +18,14 @@ using std::set;
 #ifndef API_H
 #define API_H
 
-int event_add_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
-int event_add_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
-int event_remove_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
-int event_remove_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
-int event_get_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
-int event_get_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
-int event_get_neighbors(Graph *graph, struct mg_connection *nc, uint64_t node_id);
-int event_shortest_path(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
+uint64_t event_add_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
+uint64_t event_add_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
+uint64_t event_remove_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
+uint64_t event_remove_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
+uint64_t event_get_node(Graph *graph, struct mg_connection *nc, uint64_t node_id);
+uint64_t event_get_edge(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
+uint64_t event_get_neighbors(Graph *graph, struct mg_connection *nc, uint64_t node_id);
+uint64_t event_shortest_path(Graph *graph, struct mg_connection *nc, uint64_t node_a_id, uint64_t node_b_id);
 
 
 #endif
